requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backoff==1.11.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
spotipy==2.23.0
sqlparse==0.4.4
urllib3==2.0.2
yarl==1.9.2
